Fixed changes in pr
